+++ src/filters.php
@@ -2,22 +2,6 @@
 
 namespace App;
 
-use Roots\Sage\Template;
-use Roots\Sage\Template\Wrapper;
-
-/**
- * Determine which pages should NOT display the sidebar
- * @link https://codex.wordpress.org/Conditional_Tags
- */
-add_filter('sage/display_sidebar', function ($display) {
-    // The sidebar will NOT be displayed if ANY of the following return true
-    return $display ? !in_array(true, [
-        is_404(),
-        is_front_page(),
-        is_page_template('templates/template-custom.php'),
-    ]) : $display;
-});
-
 /**
  * Add <body> classes
  */
@@ -45,11 +29,31 @@ add_filter('excerpt_more', function () {
 });
 
 /**
- * Use theme wrapper
+ * Template Hierarchy should search for .blade.php files
  */
-add_filter('template_include', function ($main) {
-    if (!is_string($main) && !(is_object($main) && method_exists($main, '__toString'))) {
-        return $main;
-    }
-    return ((new Template(new Wrapper($main)))->layout());
-}, 109);
+array_map(function ($type) {
+    add_filter("{$type}_template_hierarchy", function ($templates) {
+        return call_user_func_array('array_merge', array_map(function ($template) {
+            $normalizedTemplate = str_replace('.', '/', sage('blade')->normalizeViewPath($template));
+            return ["{$normalizedTemplate}.blade.php", "{$normalizedTemplate}.php"];
+        }, $templates));
+    });
+}, [
+    'index', '404', 'archive', 'author', 'category', 'tag', 'taxonomy', 'date', 'home',
+    'front_page', 'page', 'paged', 'search', 'single', 'singular', 'attachment'
+]);
+
+/**
+ * Render page using Blade
+ */
+add_filter('template_include', function ($template) {
+    echo template($template, apply_filters('sage/template_data', []));
+
+    // Return a blank file to make WordPress happy
+    return get_template_directory() . '/index.php';
+}, 1000);
+
+/**
+ * Tell WordPress how to find the compiled path of comments.blade.php
+ */
+add_filter('comments_template', 'App\\template_path');