+++ src/setup.php
@@ -2,7 +2,36 @@
 
 namespace App;
 
-use Roots\Sage\Template;
+use Roots\Sage\Assets\JsonManifest;
+use Roots\Sage\Template\BladeProvider;
+
+/**
+ * Add JsonManifest to Sage container
+ */
+sage()->singleton('sage.assets', function () {
+    return new JsonManifest(
+        get_stylesheet_directory().'/dist/assets.json',
+        get_stylesheet_directory_uri().'/dist'
+    );
+});
+
+/**
+ * Add Blade to Sage container
+ */
+sage()->singleton('sage.blade', function () {
+    $cachePath = wp_upload_dir()['basedir'].'/cache/compiled';
+    if (!file_exists($cachePath)) {
+        wp_mkdir_p($cachePath);
+    }
+    return new BladeProvider(TEMPLATEPATH, $cachePath, sage());
+});
+
+/**
+ * Create @asset() Blade directive
+ */
+sage('blade')->compiler()->directive('asset', function ($asset) {
+    return '<?= App\\asset_path(\''.trim($asset, '\'"').'\'); ?>';
+});
 
 /**
  * Theme assets
@@ -28,13 +57,13 @@ add_action('after_setup_theme', function () {
 
     /**
      * Enable plugins to manage the document title
-     * @link http://codex.wordpress.org/Function_Reference/add_theme_support#Title_Tag
+     * @link https://developer.wordpress.org/reference/functions/add_theme_support/#title-tag
      */
     add_theme_support('title-tag');
 
     /**
      * Register navigation menus
-     * @link http://codex.wordpress.org/Function_Reference/register_nav_menus
+     * @link https://developer.wordpress.org/reference/functions/register_nav_menus/
      */
     register_nav_menus([
         'primary_navigation' => __('Primary Navigation', 'sage')
@@ -42,21 +71,13 @@ add_action('after_setup_theme', function () {
 
     /**
      * Enable post thumbnails
-     * @link http://codex.wordpress.org/Post_Thumbnails
-     * @link http://codex.wordpress.org/Function_Reference/set_post_thumbnail_size
-     * @link http://codex.wordpress.org/Function_Reference/add_image_size
+     * @link https://developer.wordpress.org/themes/functionality/featured-images-post-thumbnails/
      */
     add_theme_support('post-thumbnails');
 
     /**
-     * Enable post formats
-     * @link http://codex.wordpress.org/Post_Formats
-     */
-    add_theme_support('post-formats', ['aside', 'gallery', 'link', 'image', 'quote', 'video', 'audio']);
-
-    /**
      * Enable HTML5 markup support
-     * @link http://codex.wordpress.org/Function_Reference/add_theme_support#HTML5
+     * @link https://developer.wordpress.org/reference/functions/add_theme_support/#html5
      */
     add_theme_support('html5', ['caption', 'comment-form', 'comment-list', 'gallery', 'search-form']);