+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### HEAD
+* Add Blade ([#1765](https://github.com/roots/sage/pull/1765))
+* Remove sidebar defaults ([#1760](https://github.com/roots/sage/pull/1760))
+* Remove post formats ([#1759](https://github.com/roots/sage/pull/1759))
+
 ### 9.0.0-alpha.4: November 16th, 2016
 * Use new webpack api schema ([8ac5f15](https://github.com/roots/sage/commit/e6e60aa))
 * Update dependencies ([70ebba7](https://github.com/roots/sage/commit/70ebba7))